+++ src/edit/qt/editor/widget/tag_editor_edit_widget.cpp
@@ -174,6 +174,8 @@ namespace Invader::EditQt {
                     textbox->setText(QString::number(std::get<std::int64_t>(current_value)));
                 }
 
+                textbox->cursorBackward(false, 32767);
+
                 value_index++;
                 return textbox;
             };